// src/generated/api.d.ts
export interface components {
  schemas: {
// ...
    TagCreateRequest: {
      /**
       * Code
       * @description Unique code for this tag
       */
      code: string;
      /**
       * Name
       * @description Tag name
       */
      name: string;
      /**
       * Description
       * @description Optional description
       */
      description?: string | null;
      /**
       * Is System Tag
       * @description True for system tags (cannot be deleted)
       * @default false
       */
      is_system_tag: boolean;
    };
    /**
     * TagListResponse
     * @description Paginated list response for Tags.
     */
    TagListResponse: {
      /**
       * Items
       * @description List of tags
       */
      items: components['schemas']['TagResponse'][];
      /**
       * Total
       * @description Total number of tags
       */
      total: number;
      /**
       * Page
       * @description Current page number
       */
      page: number;
      /**
       * Page Size
       * @description Number of items per page
       */
      page_size: number;
      /**
       * Total Pages
       * @description Total number of pages
       */
      total_pages: number;
      /**
       * Has Next
       * @description Whether there is a next page
       */
      has_next: boolean;
      /**
       * Has Previous
       * @description Whether there is a previous page
       */
      has_previous: boolean;
    };
    /**
     * TagResponse
     * @description Response model for Tag.
     */
    TagResponse: {
      /**
       * Id
       * Format: uuid
       * @description Tag ID
       */
      id: string;
      /**
       * Organisation Id
       * @description Organisation ID (None for system tags)
       */
      organisation_id?: string | null;
      /**
       * Code
       * @description Unique code for this tag
       */
      code: string;
      /**
       * Name
       * @description Tag name
       */
      name: string;
      /**
       * Description
       * @description Optional description
       */
      description?: string | null;
      /**
       * Is System Tag
       * @description True for system tags (cannot be deleted)
       */
      is_system_tag: boolean;
      /**
       * Created At
       * Format: date-time
       * @description Creation timestamp
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       * @description Last update timestamp
       */
      updated_at: string;
      /**
       * Created By
       * @description User who created this tag
       */
      created_by?: string | null;
      /**
       * Updated By
       * @description User who last updated this tag
       */
      updated_by?: string | null;
      /**
       * Is Deleted
       * @description Soft delete flag
       * @default false
       */
      is_deleted: boolean;
    };
    /**
     * TagUpdateRequest
     * @description Request model for updating a Tag.
     */
    TagUpdateRequest: {
      /**
       * Name
       * @description Tag name
       */
      name?: string | null;
      /**
       * Description
       * @description Optional description
       */
      description?: string | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}